// .PRD/erp-refactor/v3/types/payment-entry.ts
import { z } from 'zod';

export const PaymentEntrySchema = z.object({
  id: z.string(),
  owner: z.string().optional(),
  creation: z.string().optional(),
  modified: z.string().optional(),
  modified_by: z.string().optional(),
  docstatus: z.number().default(0),
  naming_series: z.enum(['ACC-PAY-.YYYY.-']),
  payment_type: z.enum(['Receive', 'Pay', 'Internal Transfer']),
  payment_order_status: z.enum(['Initiated', 'Payment Ordered']).optional(),
  posting_date: z.string().default('Today'),
  company: z.string(),
  mode_of_payment: z.string().optional(),
  party_type: z.string().optional(),
  party: z.string().optional(),
  party_name: z.string().optional(),
  book_advance_payments_in_separate_party_account: z.boolean().optional().default(false),
  reconcile_on_advance_payment_date: z.boolean().optional().default(false),
  apply_tds: z.boolean().optional().default(false),
  tax_withholding_category: z.string().optional(),
  bank_account: z.string().optional(),
  party_bank_account: z.string().optional(),
  contact_person: z.string().optional(),
  contact_email: z.string().email().optional(),
  paid_from: z.string(),
  paid_from_account_type: z.string().optional(),
  paid_from_account_currency: z.string(),
  paid_to: z.string(),
  paid_to_account_type: z.string().optional(),
  paid_to_account_currency: z.string(),
  paid_amount: z.number(),
  paid_amount_after_tax: z.number().optional(),
  source_exchange_rate: z.number(),
  base_paid_amount: z.number(),
  base_paid_amount_after_tax: z.number().optional(),
  received_amount: z.number(),
  received_amount_after_tax: z.number().optional(),
  target_exchange_rate: z.number(),
  base_received_amount: z.number(),
  base_received_amount_after_tax: z.number().optional(),
  references: z.array(z.unknown()).optional(),
  total_allocated_amount: z.number().optional(),
  base_total_allocated_amount: z.number().optional(),
  unallocated_amount: z.number().optional(),
  difference_amount: z.number().optional(),
  purchase_taxes_and_charges_template: z.string().optional(),
  sales_taxes_and_charges_template: z.string().optional(),
  taxes: z.array(z.unknown()).optional(),
  base_total_taxes_and_charges: z.number().optional(),
  total_taxes_and_charges: z.number().optional(),
  deductions: z.array(z.unknown()).optional(),
  tax_withholding_group: z.string().optional(),
  ignore_tax_withholding_threshold: z.boolean().optional().default(false),
  override_tax_withholding_entries: z.boolean().optional().default(false),
  tax_withholding_entries: z.array(z.unknown()).optional(),
  reference_no: z.string().optional(),
  reference_date: z.string().optional(),
  clearance_date: z.string().optional(),
  project: z.string().optional(),
  cost_center: z.string().optional(),
  status: z.enum(['Draft', 'Submitted', 'Cancelled']).optional().default('Draft'),
  custom_remarks: z.boolean().optional().default(false),
  remarks: z.string().optional(),
  base_in_words: z.string().optional(),
  is_opening: z.enum(['No', 'Yes']).optional().default('No'),
  letter_head: z.string().optional(),
  print_heading: z.string().optional(),
  bank: z.string().optional(),
  bank_account_no: z.string().optional(),
  payment_order: z.string().optional(),
  in_words: z.string().optional(),
  auto_repeat: z.string().optional(),
  amended_from: z.string().optional(),
  title: z.string().optional(),
});

export type PaymentEntry = z.infer<typeof PaymentEntrySchema>;

export const PaymentEntryInsertSchema = PaymentEntrySchema.omit({ id: true, owner: true, creation: true, modified: true, modified_by: true });
export type PaymentEntryInsert = z.infer<typeof PaymentEntryInsertSchema>;